refactor(tx-flow): move BatchButton component to tx-flow folder
