+++ apps/web/src/components/tx-flow/actions/Batching/index.tsx
@@ -1,13 +1,13 @@
 import { useContext, useState, type SyntheticEvent } from 'react'
 import { SafeTxContext } from '@/components/tx-flow/SafeTxProvider'
-import BatchButton from '@/components/tx/SignOrExecuteForm/BatchButton'
+import BatchButton from './BatchButton'
 import { useTxActions } from '@/components/tx/SignOrExecuteForm/hooks'
 import useIsSafeOwner from '@/hooks/useIsSafeOwner'
 import { isDelegateCall } from '@/services/tx/tx-sender/sdk'
 import { TxModalContext } from '@/components/tx-flow'
-import { TxFlowContext } from '../TxFlowProvider'
+import { TxFlowContext } from '../../TxFlowProvider'
 import useIsCounterfactualSafe from '@/features/counterfactual/hooks/useIsCounterfactualSafe'
-import { SlotName, withSlot } from '../slots'
+import { SlotName, withSlot } from '../../slots'
 import { asError } from '@safe-global/utils/services/exceptions/utils'
 import ErrorMessage from '@/components/tx/ErrorMessage'
 import { Errors, logError } from '@/services/exceptions'
